fix(api): :zap: user관련 기능이 원활히 동작되도록 코드 개선
프론트엔드에서 request시 제대로 된 response를 받을 수 없어 확인하는 도중 코드를 개선하게 되었음. 그러나 user.http에서 이미 유저관련 기능이 제대로 동작하므로 프론트엔드에서의 에러일 확률이 높음.
