+++ src/config/config.js
@@ -2,7 +2,7 @@
 require("dotenv").config();
 
 module.exports = {
-  port: process.env.PORT || 3000,
+  port: process.env.PORT || 5050,
   jwtSecret: process.env.JWT_SECRET || "your-secret-key",
   jwtExpiration: process.env.JWT_EXPIRATION || "1h",
 };
